refactor(chat): update imports and rename properties
- Change import path for CompletionOutput in AIBubble, AIBubbleFooter, and AIBubbleMetrics components.
- Rename 'timestamp' to 'createdAt' in AIBubbleFooter and UserBubble components for consistency.
- Update reasoning toggle implementation in ChatBottomBar.
- Introduce new ReasoningToggle component for better UI handling.
