+++ features/chat/components/ai-bubble.tsx
@@ -1,9 +1,9 @@
-import { useTheme } from "@/context/themes";
+import { useUserPreferences } from "@/context/user-preferences/context";
+import { ChatMessage } from "@/database/chat/types";
 import { StreamingIndicator } from "@/features/chat/components/streaming-indicator";
 import { ThinkingSection } from "@/features/chat/components/thinking-section";
-import type { ChatMessage } from "@/features/chat/model/chat-message";
-import { getAllModels } from "@/shared/ai/catalog";
-import type { CompletionOutput } from "@/shared/ai/types/runtime";
+import { getAllModels } from "@/shared/ai/text-generation/catalog";
+import { CompletionOutput } from "@/shared/ai/text-generation/types";
 import * as Clipboard from "expo-clipboard";
 import { useMemo } from "react";
 import { Text, View } from "react-native";
@@ -23,7 +23,7 @@ export function AIBubble({
   onRetry,
   isReasonEnabled,
 }: AIBubbleProps) {
-  const { colorScheme } = useTheme();
+  const { colorScheme } = useUserPreferences();
   const content = message.content ?? "";
   const reasoning = message.reasoning_content ?? "";
 
@@ -109,7 +109,7 @@ export function AIBubble({
       {!isStreaming && (
         <AIBubbleFooter
           modelDisplayName={modelDisplayName}
-          timestamp={message.timestamp}
+          createdAt={message.createdAt}
           timings={timings}
           onRetry={onRetry}
           onCopy={handleCopy}
